requirements.txt
numpy==1.24.4
opencv-python==4.10.0.84
pillow==10.4.0
openai>=1.59.8
python-dotenv==1.0.1